Add simple benchmarking framework with simple flood bench
Can be run by calling "./run" from bash.

Accompanied the comparing of the sync vs. async version of lager:dispatch_log/8
that shows the sync one is about 6.6 times slower but async one finished
logging only 3.6% of the events. Seems that there should be an option for the
user/developer: which kind the lager should be.